requirements.txt
Django==1.10.5
django-webpack-loader==0.4.1
channels==1.0.2
djangorestframework==3.5.3
django-extensions==1.7.6